inital review structure (#24)
# Add review structure

This PR aims to add review functionality to the website. This implementation allows the user to write a simple review and rate it with 0-5 stars.

## Added

This PR adds three new files
`src/product.py`
This file will contain all routes relevant to a specific product,
`src/models/review.py`
this file contains a simple object-level abstraction for a review
and `static/js/review.js` a file with functions to count stars and submit reviews to the server

## Modified

 - `src/sql/inventory/management.py`, Added method to add a new review
 - `src/sql/inventory/getters.py`, Added method to get all reviews for a given product
 - `templates/base.html` Formatting and version bump of the FA icon pack
 - `templates/product_info.html` Add everything that is needed to display/make reviews
 - `templates/jinja_macros/product_listing.html` Rewrite the URL for the clickable product
 - `src/app.py` Move product info to the product file
 - `src/models/__init__.py` add a new import

## Testing

I have tested this by creating a review, but failing to create an empty review, failing to create a review without a rating, failing to create 2 reviews from the same user on the same product, and creating reviews from two different accounts on the same product. 
 
closes #8 
closes #9